Change time param of Renderer::onUpdate() from double to float to avoid casting

+++ core/include/core/Renderer.hpp
@@ -5,6 +5,6 @@ class Renderer
 public:
     virtual bool onInit() = 0;
     virtual void onDestroy() = 0;
-    virtual void onUpdate(double time) = 0;
+    virtual void onUpdate(float time) = 0;
     virtual void onDraw() = 0;
 };
