+++ examples/02-triangle/src/TriangleRenderer.hpp
@@ -10,7 +10,7 @@ class TriangleRenderer : public Renderer
 public:
     virtual bool onInit() override;
     virtual void onDestroy() override;
-    virtual void onUpdate(double time) override {}
+    virtual void onUpdate(float time) override {}
     virtual void onDraw() override;
 private:
     ShaderProgram* m_shader = nullptr;
