+++ examples/05-texture/src/TextureRenderer.hpp
@@ -12,7 +12,7 @@ class TextureRenderer : public Renderer
 public:
     virtual bool onInit() override;
     virtual void onDestroy() override;
-    virtual void onUpdate(double time) override {}
+    virtual void onUpdate(float time) override {}
     virtual void onDraw() override;
 private:
     ShaderProgram* m_shader = nullptr;
